اتمام تمام N_Point ها

+++ src/main/java/com/example/Sanjaghak/Service/ProductAttributeService.java
@@ -1,18 +1,21 @@
 package com.example.Sanjaghak.Service;
 
+import com.example.Sanjaghak.Repository.AttributeRequirementRepository;
 import com.example.Sanjaghak.Repository.ProductAttributeRepository;
+import com.example.Sanjaghak.Repository.ProductAttributeValueRepository;
 import com.example.Sanjaghak.Repository.UserAccountsRepository;
+import com.example.Sanjaghak.Specification.AttributeSpecifications;
 import com.example.Sanjaghak.model.ProductAttribute;
-import com.example.Sanjaghak.model.Products;
 import com.example.Sanjaghak.model.UserAccounts;
 import com.example.Sanjaghak.security.jwt.JwtUtil;
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
-import java.util.List;
 import java.util.UUID;
 
 @Service
@@ -22,6 +25,12 @@ public class ProductAttributeService {
 
     @Autowired
     private UserAccountsRepository userAccountsRepository;
+
+    @Autowired
+    private ProductAttributeValueRepository productAttributeValueRepository;
+
+    @Autowired
+    private AttributeRequirementRepository attributeRequirementRepository;
 
 
     @Transactional
@@ -64,11 +73,12 @@ public class ProductAttributeService {
 
     public ProductAttribute getProductAttributeById(UUID attributeId) {
         return productAttributeRepository.findById(attributeId)
-                .orElseThrow(()-> new EntityNotFoundException("محصول مورد نظر پیدا نشد !"));
+                .orElseThrow(()-> new EntityNotFoundException("ویژگی مورد نظر پیدا نشد !"));
     }
 
-    public List<ProductAttribute> getAllProductAttributes() {
-        return productAttributeRepository.findAll();
+    public Page<ProductAttribute> getAllProductAttributes(String attributeName, Pageable pageable) {
+        return productAttributeRepository.findAll(
+                AttributeSpecifications.filterAttribute(attributeName), pageable);
     }
 
     public void deleteProductAttribute(UUID attributeId,String token) {
@@ -77,9 +87,17 @@ public class ProductAttributeService {
         if (!role.equalsIgnoreCase("admin") && !role.equalsIgnoreCase("manager")) {
             throw new RuntimeException("شما مجوز لازم برای انجام این عملیات را ندارید");
         }
+
         if (!productAttributeRepository.existsById(attributeId)) {
             throw new IllegalArgumentException("محصول مورد نظر یافت نشد.");
         }
+        boolean isUsed = productAttributeValueRepository.existsByAttributeId_attributeId(attributeId) || attributeRequirementRepository.existsByAttributeId_attributeId(attributeId);
+
+
+        if (isUsed) {
+            throw new IllegalStateException("این ویژگی در بخش‌های دیگر استفاده شده و قابل حذف نیست");
+        }
+
         ProductAttribute delete = productAttributeRepository.findById(attributeId).orElseThrow((() -> new EntityNotFoundException("محصول مورد نظر پیدا نشد !")));
         productAttributeRepository.delete(delete);
